make linters happy

+++ gz_ros2_control/src/gz_ros2_control_plugin.cpp
@@ -212,8 +212,7 @@ GazeboSimROS2ControlPluginPrivate::GetEnabledJoints(
         {
           RCLCPP_INFO(
             node_->get_logger(), "[gz_ros2_control] Fixed joint ['%s'] (Entity='%lu') is skipped.",
-            jointName.c_str(), static_cast<unsigned long>(jointEntity)
-          );
+            jointName.c_str(), jointEntity);
           continue;
         }
       case sdf::JointType::REVOLUTE2:
@@ -225,8 +224,7 @@ GazeboSimROS2ControlPluginPrivate::GetEnabledJoints(
             node_->get_logger(),
             "[gz_ros2_control] Joint ['%s'] (Entity='%lu') is of unsupported type."
             " Only joints with a single axis are supported.",
-            jointName.c_str(), static_cast<unsigned long>(jointEntity)
-          );
+            jointName.c_str(), jointEntity);
           continue;
         }
       default:
